#699  Move RequestData and Context classes to relevant modules
